requirements.txt
requests
python-dotenv
selenium
webdriver-manager
sentence-transformers
scikit-learn
nltk
opencv-python
numpy